sx119: milestone of CNN visualize, display to be improved

+++ src/main/kotlin/processing/filters/CNNVisualization.kt
@@ -14,48 +14,47 @@ import java.io.File
 
 @Serializable
 @SerialName("CNNVisualize")
-class CNNVisualization(val netName: String, val layerNum: Int, val channelNum: List<Int>, val height: Int, val width: Int): ImageProcessing {
+class CNNVisualization(
+    val netName: String,
+    val imgShape: List<Int>,
+    val layerNum: Int,
+    val channelNum: List<Int>): ImageProcessing {
 
     @Transient
     var net: List<String> = listOf()
 
     override fun process(srcImage: WritableImage, destImage: WritableImage) {
-        if (!checkLoaded(netName)) {
-            println("loading new net")
-            val exitCode = loadNet(netName)
-            println("loading net finish with exit code $exitCode")
-        }
+        val srcHeight = srcImage.height.toInt()
+        val srcWidth = srcImage.width.toInt()
+        loadNet(netName, imgShape)
         net = getLayerPaths()
 
-        // TODO: use user self defined preprocess method
         val reader = srcImage.pixelReader
-        val h = srcImage.height.toInt()
-        val w = srcImage.width.toInt()
         val pixels = Array(3) {
-            Array(w) {
+            Array(srcWidth) {
                 DoubleArray(
-                    h
+                    srcHeight
                 )
             }
         }
-        for (i in 0 until w) {
-            for (j in 0 until h) {
+        for (i in 0 until srcWidth) {
+            for (j in 0 until srcHeight) {
                 pixels[0][i][j] = reader.getColor(i, j).red
                 pixels[1][i][j] = reader.getColor(i, j).green
                 pixels[2][i][j] = reader.getColor(i, j).blue
             }
         }
-        val dimension = 3 * h * w
+        val dimension = 3 * srcHeight * srcWidth
         val buf = FloatArray(dimension)
-        for (i in 0 until w) {
-            for (j in 0 until h) {
-                buf[j * w + i] = pixels[0][i][j].toFloat()
-                buf[j * w + i + h*w] = pixels[1][i][j].toFloat()
-                buf[j * w + i + h*w*2] = pixels[2][i][j].toFloat()
+        for (i in 0 until srcWidth) {
+            for (j in 0 until srcHeight) {
+                buf[j * srcWidth + i] = pixels[0][i][j].toFloat()
+                buf[j * srcWidth + i + srcHeight*srcWidth] = pixels[1][i][j].toFloat()
+                buf[j * srcWidth + i + srcHeight*srcWidth*2] = pixels[2][i][j].toFloat()
             }
         }
 
-        var data = IValue.from(Tensor.fromBlob(buf, longArrayOf(1, 3, h.toLong(), w.toLong())))
+        var data = IValue.from(Tensor.fromBlob(buf, longArrayOf(1, 3, srcHeight.toLong(), srcWidth.toLong())))
         var layerCnt = 0
         for (path in net) {
             val layer = Module.load(path)
@@ -66,18 +65,23 @@ class CNNVisualization(val netName: String, val layerNum: Int, val channelNum: L
             layerCnt ++
         }
 
-        val outputImage = WritableImage(width, height)
+        val metadata = File("./src/main/resources/CNN_split/CNN_traced/.Metadata.log").readLines()
+        val tokens = metadata[layerNum + 1].split('|') // +1 for the first line of metadata is not layer information
+        val outputWidth = tokens[tokens.size-1].toInt()
+        val outputHeight = tokens[tokens.size-2].toInt()
+        val outputImage = WritableImage(outputWidth, outputHeight)
         val writer = outputImage.pixelWriter
         val output = data.toTensor().dataAsFloatArray
-        val dimensionIndex = channelNum.reduce { x, y -> x * y }
-        for (i in 0 until width) {
+//        val dimensionIndex = channelNum.reduce { x, y -> x * y }
+        val dimensionIndex = 0
+        for (i in 0 until outputWidth) {
             val min =
-                output.copyOfRange(dimensionIndex * width * height, (dimensionIndex + 1) * width * height).minOrNull()!!
+                output.copyOfRange(dimensionIndex * outputWidth * outputHeight, (dimensionIndex + 1) * outputWidth * outputHeight).minOrNull()!!
             val range =
-                output.copyOfRange(dimensionIndex * width * height, (dimensionIndex + 1) * width * height).maxOrNull()!! - min
+                output.copyOfRange(dimensionIndex * outputWidth * outputHeight, (dimensionIndex + 1) * outputWidth * outputHeight).maxOrNull()!! - min
 
-            for (j in 0 until height) {
-                val pixelVal = output[dimensionIndex * width * height + j * width + i].toDouble()
+            for (j in 0 until outputHeight) {
+                val pixelVal = output[dimensionIndex * outputWidth * outputHeight + j * outputWidth + i].toDouble()
                 val castedVal = ((pixelVal - min) / range).coerceIn(0.0, 1.0)
                 val color = Color.color(
                     castedVal,
@@ -86,18 +90,26 @@ class CNNVisualization(val netName: String, val layerNum: Int, val channelNum: L
                 writer.setColor(i, j, color)
             }
         }
+//
+//        for (i in 0 until outputHeight) {
+//            for (j in 0 until 3) {
+//                print(output[dimensionIndex * outputWidth * outputHeight + i * outputWidth + j].toString() + ' ')
+//            }
+//            print("...")
+//            for (j in outputWidth - 3 until outputWidth)
+//                print(output[dimensionIndex * outputWidth * outputHeight + i * outputWidth + j].toString() + ' ')
+//            println()
+//        }
 
         // stretch to dst image size
         val view = ImageView(outputImage)
-        view.fitWidth = w.toDouble()
-        view.fitHeight = h.toDouble()
+        view.fitWidth = srcWidth.toDouble()
+        view.fitHeight = srcHeight.toDouble()
         val stretchedImage = view.snapshot(null, null)
-        print("streched image: " + stretchedImage.width.toString() + ' ' + stretchedImage.height)
-        print("expected: " + w + " " + h)
         val outputReader = stretchedImage.pixelReader
         val dstWriter = destImage.pixelWriter
-        for (i in 0 until w)
-            for (j in 0 until h)
+        for (i in 0 until srcWidth)
+            for (j in 0 until srcHeight)
                 dstWriter.setColor(i, j, outputReader.getColor(i, j))
     }
 
@@ -110,11 +122,12 @@ class CNNVisualization(val netName: String, val layerNum: Int, val channelNum: L
          * check if net loaded in directory 'resources/CNN_split/CNN_traced' is same net as expectedNet
          *
          * @param expectedNet: path to expectedNet loaded in directory
+         * @param expectedNet: path to expectedNet loaded in directory
          */
-        fun checkLoaded(expectedNet: String): Boolean {
+        fun checkLoaded(expectedNet: String, expectedShape: List<Int>): Boolean {
             val lines = File("./src/main/resources/CNN_split/CNN_traced/.Metadata.log")
                 .readLines()
-            return lines[0] == expectedNet
+            return lines[0] == expectedNet + ' ' + expectedShape.joinToString()
         }
 
         /**
@@ -139,11 +152,14 @@ class CNNVisualization(val netName: String, val layerNum: Int, val channelNum: L
          * @param path: path of loaded CNN
          * @return: exit code of loading python script
          */
-        fun loadNet(path: String): Int {
+        fun loadNet(path: String, shape: List<Int>): Int {
+            if (checkLoaded(path, shape))
+                return 0
+
             val removeProcess = Runtime.getRuntime().exec("rm -r -f ./src/main/resources/CNN_split/CNN_traced")
             removeProcess.waitFor()
 
-            val process = Runtime.getRuntime().exec("python3 ./src/main/resources/CNN_split/CNN_spliter.py $path ")
+            val process = Runtime.getRuntime().exec("python3 ./src/main/resources/CNN_split/CNN_spliter.py $path ${shape.joinToString(separator = ",")}")
             return process.waitFor()
         }
     }
